refactor(store): deepen durability publication
Centralize temp publication while moving global maintenance to its domain owners.
