+++ crates/rimz/src/store/AGENTS.md
@@ -7,7 +7,7 @@ Local contract for `crates/rimz/src/store/` — durable workspace state. Extends
 - Every mutator and its public intent/outcome vocabulary lands under [`writer.rs`](./writer.rs): the façade owns the ordinary commit and log-boundary primitives, and `writer/` owns the debounce, lifecycle, publish, reset, queue, and reap branches. Reads on the `Store` handle stay lock-free; snapshot schema is consumed through its owning module.
 - `workspace.lock` is the only cross-process serialization. Every writer is a short-lived CLI process, so a change that needs ordering takes the lock rather than introducing an in-process actor.
 - Every durable write routes through [`atomic.rs`](./atomic.rs), and every fsync syscall lives inside it; `cargo xtask invariants` rejects a `sync_all`/`sync_data` call anywhere else. No module hand-rolls its own atomic dance, and the event-log frame encoding stays beside its decoder in [`event_log/frame.rs`](./event_log/frame.rs).
-- [`writer/queue.rs`](./writer/queue.rs) owns all three message surfaces together: the live `messages.jsonl`, terminal text in the sibling `history.jsonl`, and terminal outcomes in the event log. Shared age pruning stays in [`atomic.rs`](./atomic.rs).
+- [`writer/queue.rs`](./writer/queue.rs) owns all three message surfaces together: the live `messages.jsonl`, terminal text in the sibling `history.jsonl`, and terminal outcomes in the event log. Archive retention stays with event-log rotation, global orphan-temp collection with GC, and cache-local temp-sibling cleanup with [`atomic.rs`](./atomic.rs).
 - The group sync and the checkpoint publish debounce through stamps beside the workspace lock, which is what keeps the write path O(1) over log history. Wakeups fire before the publish: consumers fold the log tail from their own cursor, so checkpoint cadence is latency tuning and never truth.
 
 ## Reads
