+++ crates/rimz/src/store/event_log/rotation.rs
@@ -1,7 +1,7 @@
 use std::fs;
 use std::io;
 use std::path::{Path, PathBuf};
-use std::time::Duration;
+use std::time::{Duration, SystemTime};
 
 use uuid::Uuid;
 
@@ -18,6 +18,12 @@ pub enum RotationOutcome {
         archive_path: PathBuf,
         bytes_rotated: u64,
     },
+}
+
+#[derive(Clone, Debug, Default, PartialEq, Eq)]
+pub struct PruneOutcome {
+    pub files_removed: usize,
+    pub bytes_removed: u64,
 }
 
 impl RotationOutcome {
@@ -87,12 +93,60 @@ pub fn rotate(events_log: &Path, archive_dir: &Path, min_bytes: u64) -> Result<R
 /// considered; foreign files are ignored so a misconfigured operator can't
 /// lose data by dropping unrelated content into the archive directory.
 #[must_use = "maintenance report; surface it to the caller"]
-pub fn prune_archive(archive_dir: &Path, older_than: Duration) -> Result<atomic::PruneOutcome> {
-    Ok(atomic::prune_old_files(
-        archive_dir,
-        older_than,
-        is_archive_name,
-    )?)
+pub fn prune_archive(archive_dir: &Path, older_than: Duration) -> Result<PruneOutcome> {
+    Ok(prune_old_files(archive_dir, older_than)?)
+}
+
+fn prune_old_files(dir: &Path, older_than: Duration) -> atomic::Result<PruneOutcome> {
+    let entries = match fs::read_dir(dir) {
+        Ok(entries) => entries,
+        Err(source) if source.kind() == io::ErrorKind::NotFound => {
+            return Ok(PruneOutcome::default());
+        }
+        Err(source) => {
+            return Err(atomic::AtomicErr::Io {
+                path: dir.to_path_buf(),
+                source,
+            });
+        }
+    };
+
+    let now = SystemTime::now();
+    let mut report = PruneOutcome::default();
+    for entry in entries {
+        let entry = entry.map_err(|source| atomic::AtomicErr::Io {
+            path: dir.to_path_buf(),
+            source,
+        })?;
+        let path = entry.path();
+        if !is_archive_name(&path) {
+            continue;
+        }
+        let metadata = fs::symlink_metadata(&path).map_err(|source| atomic::AtomicErr::Io {
+            path: path.clone(),
+            source,
+        })?;
+        let modified = metadata
+            .modified()
+            .map_err(|source| atomic::AtomicErr::Io {
+                path: path.clone(),
+                source,
+            })?;
+        let Ok(age) = now.duration_since(modified) else {
+            continue;
+        };
+        if age < older_than {
+            continue;
+        }
+        let bytes = metadata.len();
+        fs::remove_file(&path).map_err(|source| atomic::AtomicErr::Io {
+            path: path.clone(),
+            source,
+        })?;
+        report.files_removed += 1;
+        report.bytes_removed = report.bytes_removed.saturating_add(bytes);
+    }
+    Ok(report)
 }
 
 fn is_archive_name(path: &Path) -> bool {
